// src/redux-middleware/types.ts
export const enum ActionTypes {
  READ_TIMELINE = 'usecases/READ_TIMELINE',
  FINISHED_READING_TIMELINE = 'usecases/FINISHED_READING_TIMELINE',
  FAILED_TO_READ_TIMELINE = 'usecases/FAILED_TO_READ_TIMELINE',
  CREATE_USER = 'usecases/CREATE_USER',
  FINISHED_CREATING_USER = 'usecases/FNISHED_CREATING_USER',
  FAILED_TO_CREATE_USER = 'usecases/FAILED_TO_CREATE_USER',
  TWEET = 'usecases/TWEET',
  FINISHED_TWEETING = 'usecases/FINISHED_TWHEETING',
  FAILED_TO_TWEET = 'usecases/FAILED_TO_TWEET',
  RETWEET = 'usecases/RETWEET',
  FINISHED_RETWEETING = 'usecases/FINISHED_RETWHEETING',
  FAILED_TO_RETWEET = 'usecases/FAILED_TO_RETWEET'
}